gdw_data_core/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gdw-data-core"
version = "0.1.0"
description = "Reusable framework for data migration pipelines"
authors = [{name = "GDW Data Architect"}]
dependencies = [
    "apache-beam[gcp]==2.49.0",
    "apache-airflow>=2.5.0",
    "google-cloud-pubsub==2.18.0",
    "google-cloud-storage==2.10.0",
    "google-cloud-bigquery==3.12.0",
    "pydantic>=2.0.0",
]

[tool.setuptools]
packages = ["gdw_data_core"]
package-dir = {"gdw_data_core" = "."}
